booking-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
psycopg2-binary==2.9.10
redis==7.4.0
kafka-python==2.3.1
httpx==0.28.1
pydantic==2.11.5
python-dotenv==1.2.2
